+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@
 - The new parameter '`RecipientProperties`' controls which recipient properties are loaded and can be used in '`GrantorFilter`' and '`TrusteeFilter`'. It also helps keep network traffic and memory usage low. See '`README`' for details.
 - Mail-enabled public folders are now considered when exporting Send As and Send On Behalf permissions
 - Support for export of public folder permissions. See '`README`' for details regarding the new parameters '`ExportPublicFolderPermissions`', '`ExportPublicFolderPermissionsAnonymous`', '`ExportPublicFolderPermissionsDefault`' and '`ExportPublicFolderPermissionsExcludeFoldertype`'.
+- Support for export management role group permissiones. See '`README`' for new parameter '`ExportManagementRoleGroupMembers`'.
 ### Fixed
 - Export all Send As permissions, not only the one granted by the last recipient checked by each parallel job